separated cases into three components
