+++ vite.config.js
@@ -1,3 +1,6 @@
 module.exports = {
   port: 8080,
+  vueCompilerOptions: {
+    isCustomElement: (tag) => tag.startsWith('bwc-'),
+  },
 }
